+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- cli: When the end timestamp is date-only and its time precedes the start, adjust it to the
+  following day.
+
 ## [v0.5.0] - 2025-08-15
 
 ### Added