+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.5.1] - 2025-08-19
+
 ### Changed
 
 - cli: Show all events in today by default
@@ -205,7 +207,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add events and todos command
 - Add dashboard command
 
-[unreleased]: https://github.com/yzx9/aim/compare/v0.5.0...HEAD
+[unreleased]: https://github.com/yzx9/aim/compare/v0.5.1...HEAD
+[v0.5.1]: https://github.com/yzx9/aim/compare/v0.5.0...v0.5.1
 [v0.5.0]: https://github.com/yzx9/aim/compare/v0.4.0...v0.5.0
 [v0.4.0]: https://github.com/yzx9/aim/compare/v0.3.2...v0.4.0
 [v0.3.2]: https://github.com/yzx9/aim/compare/v0.3.1...v0.3.2